% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empaticaR-package.R
\docType{package}
\name{empaticaR-package}
\alias{empaticaR}
\alias{empaticaR-package}
\title{empaticaR: Read and organize data from Empatica EmbracePlus watch}
\description{
Read and organize data from Empatica EmbracePlus watch.
}
\author{
\strong{Maintainer}: Brian Lau \email{brian.lau@upmc.fr} (\href{https://orcid.org/0000-0001-8615-8244}{ORCID})

}
\keyword{internal}
