% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata.R
\name{read_empatica_avro_dir}
\alias{read_empatica_avro_dir}
\title{read_empatica_avro_dir: read multiple AVRO files in directory}
\usage{
read_empatica_avro_dir(path, sc = NULL)
}
\arguments{
\item{path}{Path of directory}

\item{sc}{A Spark connection}
}
\value{
A nested data frame with a row for each AVRO file
}
\description{
read_empatica_avro_dir: read multiple AVRO files in directory
}
\examples{
basepath = "/Users/brian/Downloads/2023-11-27/01-3YK3J151FF/raw_data/v6/"
sc <- sparklyr::spark_connect(master = "local", version = "3.5.0", packages = c("avro"))
df = read_empatica_avro_dir(sc = sc, path = basepath)
}
