% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata.R
\name{read_empatica_avro_file}
\alias{read_empatica_avro_file}
\title{read_empatica_avro_file: Read single AVRO file}
\usage{
read_empatica_avro_file(
  path,
  sc = NULL,
  spark_version = "3.5.0",
  spark_disconnect_on_exit = FALSE,
  collect_on_exit = TRUE,
  parse_rawdata = TRUE,
  ...
)
}
\arguments{
\item{path}{Path to file}

\item{sc}{A Spark connection}

\item{...}{}
}
\value{
A nested data frame
}
\description{
read_empatica_avro_file: Read single AVRO file
}
\examples{
basepath = "/Users/brian/Downloads/2023-11-27/01-3YK3J151FF/raw_data/v6/"
sc <- spark_connect(master = "local", version = "3.5.0", packages = c("avro"))
df = read_empatica_avro_file(sc = sc, path = paste0(basepath, "1-1-01_1701129261.avro"))
}
